+++ BE/src/main/java/com/oss/maeumnaru/diary/entity/DiaryAnalysisEntity.java
@@ -23,7 +23,7 @@ public class DiaryAnalysisEntity {
 
     private String createDate;
 
-    private long emotionRate;
+    private float emotionRate;
 
     private long mealCount;
